updated UserAPIDatastore
